+++ packages/cli/src/cmds/agentInstaller/projectConfiguration.ts
@@ -1,10 +1,15 @@
+import assert from 'assert';
 import chalk from 'chalk';
 import { promises as fs, constants as fsConstants } from 'fs';
+
 import { basename, join, resolve } from 'path';
 import { AbortError, InvalidPathError } from '../errors';
 import UI from '../userInteraction';
 import AgentInstaller from './agentInstaller';
 import getAvailableInstallers from './installers';
+import { YarnInstaller } from './javaScriptAgentInstaller';
+import CommandStruct from './commandStruct';
+import { run } from './commandRunner';
 
 export interface ProjectConfiguration {
   name: string;
@@ -81,10 +86,110 @@ async function resolveSelectedInstallers(
     }
   }
 
-  // Sanity check. Verify everything resolved properly. This should branch should never occur.
+  // Sanity check. Verify everything resolved properly. This branch should never occur.
   if (projects.some(({ selectedInstaller }) => !selectedInstaller)) {
     throw new Error('Invalid selection');
   }
+}
+
+function yarn1Workspaces(info: unknown): [string, string][] {
+  assert(info && typeof info === 'object');
+  return Object.entries(info).map(([name, data]: [string, unknown]) => {
+    assert(data && typeof data === 'object');
+    assert('location' in data);
+    assert(typeof data['location'] === 'string');
+    return [name, data['location']];
+  });
+}
+// With yarn 1:
+// - if there's no "workspaces" key in package.json we get the following error:
+//   $ yarn workspaces info --json
+//   yarn workspaces v1.22.19
+//   error Cannot find the root of your workspace - are you sure you're currently in a workspace?
+//   info Visit https://yarnpkg.com/en/docs/cli/workspaces for documentation about this command.
+// - if there is a "workspaces" key in package.json the output contains text
+//   outside of json. Remove it:
+//   $ yarn workspaces info --json
+//   yarn workspaces v1.22.19
+//   {
+//    "workspace-a": {
+//      "location": "workspace-a",
+//      "workspaceDependencies": [],
+//      "mismatchedWorkspaceDependencies": []
+//    }
+//   }
+//   Done in 0.02s.
+async function getYarnSubprojectsVersionOne(
+  dir: string,
+  installer: YarnInstaller,
+  subprojects: ProjectConfiguration[]
+) {
+  const cmd = new CommandStruct('yarn', ['workspaces', 'info', '--json'], installer.path);
+  const output = await run(cmd);
+
+  // Remove any text before and after JSON
+  let lines = output.stdout.split('\n');
+  let outputClean = '';
+  for (const line of lines) {
+    if (line.startsWith('error')) return;
+    if (!(line.startsWith('yarn workspaces v') || line.startsWith('Done in '))) outputClean += line;
+  }
+
+  for (const [name, location] of yarn1Workspaces(JSON.parse(outputClean))) {
+    const subProjectPath = join(dir, location);
+    const subProjectName = basename(name);
+    const projectConfiguration: ProjectConfiguration = {
+      name: subProjectName,
+      path: subProjectPath,
+      // must set the new path in the installer
+      availableInstallers: [new YarnInstaller(subProjectPath)],
+    };
+    subprojects.push(projectConfiguration);
+  }
+}
+
+async function getYarnSubprojectsVersionAboveOne(
+  dir: string,
+  installer: YarnInstaller,
+  subprojects: ProjectConfiguration[]
+) {
+  const cmd = new CommandStruct('yarn', ['workspaces', 'list', '--json'], installer.path);
+  const output = await run(cmd);
+
+  for (const line of output.stdout.split('\n')) {
+    if (line !== '') {
+      const yarnWorkspace = JSON.parse(line);
+      const subProjectPath = join(dir, yarnWorkspace['location']);
+      const subProjectName = basename(yarnWorkspace['name']);
+      const projectConfiguration: ProjectConfiguration = {
+        name: subProjectName,
+        path: subProjectPath,
+        // must set the new path in the installer
+        availableInstallers: [new YarnInstaller(subProjectPath)],
+      };
+      subprojects.push(projectConfiguration);
+    }
+  }
+}
+
+export async function getYarnSubprojects(
+  dir: string,
+  availableInstallers: AgentInstaller[]
+): Promise<ProjectConfiguration[]> {
+  let subprojects: ProjectConfiguration[] = [];
+
+  for (const installer of availableInstallers) {
+    if (installer.name === 'yarn') {
+      const yarnInstaller = installer as YarnInstaller;
+      if (await yarnInstaller.isYarnVersionOne()) {
+        await getYarnSubprojectsVersionOne(dir, yarnInstaller, subprojects);
+      } else {
+        await getYarnSubprojectsVersionAboveOne(dir, yarnInstaller, subprojects);
+      }
+    }
+  }
+
+  return subprojects;
 }
 
 /**
@@ -98,37 +203,50 @@ async function resolveSelectedInstallers(
  */
 async function getSubprojects(
   dir: string,
-  rootHasInstaller: boolean
+  rootHasInstaller: boolean,
+  availableInstallers: AgentInstaller[]
 ): Promise<ProjectConfiguration[] | undefined> {
-  const ents = await fs.readdir(dir, { withFileTypes: true });
-  const subprojects = (
-    await Promise.all(
-      ents.map(async (ent): Promise<ProjectConfiguration | null> => {
-        if (!ent.isDirectory()) {
-          return null;
-        }
+  let subprojects: ProjectConfiguration[] = [];
+  const yarnSubprojects: ProjectConfiguration[] = await getYarnSubprojects(
+    dir,
+    availableInstallers
+  );
 
-        const subProjectPath = join(dir, ent.name);
-        try {
-          await fs.access(subProjectPath, fsConstants.R_OK);
-        } catch {
-          // This directory is not readable. Ignore it.
-          return null;
-        }
+  if (yarnSubprojects.length > 0) {
+    // detect yarn packages as sub-projects
+    subprojects = yarnSubprojects;
+  } else {
+    // read directory to detect sub-projects
+    const ents = await fs.readdir(dir, { withFileTypes: true });
+    subprojects = (
+      await Promise.all(
+        ents.map(async (ent): Promise<ProjectConfiguration | null> => {
+          if (!ent.isDirectory()) {
+            return null;
+          }
 
-        const installers = await getAvailableInstallers(subProjectPath);
-        if (!installers.length) {
-          return null;
-        }
+          const subProjectPath = join(dir, ent.name);
+          try {
+            await fs.access(subProjectPath, fsConstants.R_OK);
+          } catch {
+            // This directory is not readable. Ignore it.
+            return null;
+          }
 
-        return {
-          name: ent.name,
-          path: subProjectPath,
-          availableInstallers: installers,
-        };
-      })
-    )
-  ).filter(Boolean) as ProjectConfiguration[];
+          const installers = await getAvailableInstallers(subProjectPath);
+          if (!installers.length) {
+            return null;
+          }
+
+          return {
+            name: ent.name,
+            path: subProjectPath,
+            availableInstallers: installers,
+          };
+        })
+      )
+    ).filter(Boolean) as ProjectConfiguration[];
+  }
 
   if (!subprojects.length) {
     // There's nothing that we can do here.
@@ -143,7 +261,7 @@ async function getSubprojects(
       'This directory contains sub-projects. Would you like to choose sub-projects for installation?',
   });
   if (!addSubprojects) {
-    // The user has opted not to continue by selecting subprojects.
+    // The user has opted not to continue by not selecting subprojects.
     return undefined;
   }
 
@@ -175,7 +293,7 @@ export async function getProjects(
   const availableInstallers = await getAvailableInstallers(dir);
   const rootHasInstaller = !!availableInstallers.length;
   if (allowMulti) {
-    const subprojects = await getSubprojects(dir, rootHasInstaller);
+    const subprojects = await getSubprojects(dir, rootHasInstaller, availableInstallers);
     if (subprojects?.length) {
       projectConfigurations.push(...(subprojects as ProjectConfiguration[]));
     } else if (subprojects?.length === 0) {