+++ packages/cli/tests/unit/agentInstall/installCommand.spec.ts
@@ -1,4 +1,4 @@
-import path from 'path';
+import path, { join } from 'path';
 import fse from 'fs-extra';
 import tmp from 'tmp';
 import sinon, { SinonStub } from 'sinon';
@@ -12,6 +12,7 @@ import AgentInstallerProcedure from '../../../src/cmds/agentInstaller/agentInsta
 import * as commandRunner from '../../../src/cmds/agentInstaller/commandRunner';
 import CommandStruct, { CommandReturn } from '../../../src/cmds/agentInstaller/commandStruct';
 import * as ProjectConfiguration from '../../../src/cmds/agentInstaller/projectConfiguration';
+import { getYarnSubprojects } from '../../../src/cmds/agentInstaller/projectConfiguration';
 
 import UI from '../../../src/cmds/userInteraction';
 import { ValidationError } from '../../../src/cmds/errors';
@@ -25,6 +26,7 @@ import { dump } from 'js-yaml';
 
 import * as openTicket from '../../../src/lib/ticket/openTicket';
 import { withStubbedTelemetry } from '../../helper';
+import { YarnInstaller } from '../../../src/cmds/agentInstaller/javaScriptAgentInstaller';
 
 const fixtureDir = path.join(__dirname, '..', 'fixtures');
 tmp.setGracefulCleanup();
@@ -749,7 +751,9 @@ appmap_dir: tmp/appmap
 
   describe('A JavaScript project', () => {
     const jsTestE2E = async (
-      nodeVersion: (command: CommandStruct) => Promise<CommandReturn>,
+      installerName: string,
+      getVersion: (command: CommandStruct) => Promise<CommandReturn>,
+      getYarnWorkspaces: (command: CommandStruct) => Promise<CommandReturn>,
       installAgent: (command: CommandStruct) => Promise<CommandReturn>,
       initAgent: (command: CommandStruct) => Promise<CommandReturn>,
       validateAgent: (command: CommandStruct) => Promise<CommandReturn>,
@@ -757,19 +761,40 @@ appmap_dir: tmp/appmap
       checkCurrentConfig: (command: CommandStruct) => Promise<CommandReturn>
     ) => {
       let callIdx = 0;
-      sinon
-        .stub(commandRunner, 'run')
-        .onCall(callIdx++)
-        .callsFake(nodeVersion)
-        .onCall(callIdx++)
-        .callsFake(checkCurrentConfig)
-        .onCall(callIdx++)
-        .callsFake(installAgent)
-        .onCall(callIdx++)
-        .callsFake(initAgent)
-        .onCall(callIdx++)
-        .callsFake(validateAgent)
-        .onCall(callIdx++);
+      if (installerName === 'yarn') {
+        sinon
+          .stub(commandRunner, 'run')
+          .onCall(callIdx++)
+          .callsFake(getVersion)
+          .onCall(callIdx++)
+          .callsFake(getYarnWorkspaces)
+          .onCall(callIdx++)
+          // why must installAgent be here and twice for the yarn test to pass?
+          .callsFake(installAgent)
+          .onCall(callIdx++)
+          .callsFake(checkCurrentConfig)
+          .onCall(callIdx++)
+          .callsFake(installAgent)
+          .onCall(callIdx++)
+          .callsFake(initAgent)
+          .onCall(callIdx++)
+          .callsFake(validateAgent)
+          .onCall(callIdx++);
+      } else {
+        sinon
+          .stub(commandRunner, 'run')
+          .onCall(callIdx++)
+          .callsFake(getVersion)
+          .onCall(callIdx++)
+          .callsFake(checkCurrentConfig)
+          .onCall(callIdx++)
+          .callsFake(installAgent)
+          .onCall(callIdx++)
+          .callsFake(initAgent)
+          .onCall(callIdx++)
+          .callsFake(validateAgent)
+          .onCall(callIdx++);
+      }
 
       return invokeCommand(projectDir, evalResults);
     };
@@ -791,6 +816,28 @@ packages:
 language: javascript
 appmap_dir: tmp/appmap
 `;
+
+    const yarnVersion = (cmdStruct: CommandStruct) => {
+      expect(cmdStruct.program).toEqual('yarn');
+      const args = cmdStruct.args;
+      expect(args).toEqual(['--version']);
+      const ret = {
+        stdout: '3.2.1',
+        stderr: '',
+      };
+      return Promise.resolve(ret);
+    };
+
+    const getYarnWorkspaces = (cmdStruct: CommandStruct) => {
+      expect(cmdStruct.program).toEqual('yarn');
+      const args = cmdStruct.args;
+      expect(args).toEqual(['workspaces', 'list', '--json']);
+      const ret = {
+        stdout: '',
+        stderr: '',
+      };
+      return Promise.resolve(ret);
+    };
 
     const initAgent = (cmdStruct: CommandStruct) => {
       expect(cmdStruct.program).toEqual('npx');
@@ -849,7 +896,9 @@ appmap_dir: tmp/appmap
         };
 
         await jsTestE2E(
+          'npm',
           nodeVersion,
+          getYarnWorkspaces,
           installAgent,
           initAgent,
           validateAgent,
@@ -866,7 +915,9 @@ appmap_dir: tmp/appmap
           expect(err?.message).toMatch(msg);
         };
         await jsTestE2E(
+          'npm',
           nodeVersion,
+          getYarnWorkspaces,
           installAgent,
           initAgent,
           failValidate,
@@ -875,6 +926,7 @@ appmap_dir: tmp/appmap
         );
       });
     });
+
     describe('managed with yarn', () => {
       const projectFixture = path.join(fixtureDir, 'javascript', 'yarn');
 
@@ -899,7 +951,7 @@ appmap_dir: tmp/appmap
       };
 
       it('installs as expected', async () => {
-        expect.assertions(12);
+        expect.assertions(15);
         const evalResults = (err: Error | undefined, argv: any, output: string) => {
           expect(err).toBeNull();
 
@@ -910,7 +962,9 @@ appmap_dir: tmp/appmap
         };
 
         await jsTestE2E(
-          nodeVersion,
+          'yarn',
+          yarnVersion,
+          getYarnWorkspaces,
           installAgent,
           initAgent,
           validateAgent,
@@ -920,14 +974,16 @@ appmap_dir: tmp/appmap
       });
 
       it('fails when validation fails', async () => {
-        expect.assertions(9);
+        expect.assertions(12);
         const msg = 'failValidate, validation failed';
         const failValidate = () => Promise.reject(new Error(msg));
         const evalResults = (err: Error | undefined, argv: any, output: string) => {
           expect(err?.message).toMatch(msg);
         };
         await jsTestE2E(
-          nodeVersion,
+          'yarn',
+          yarnVersion,
+          getYarnWorkspaces,
           installAgent,
           initAgent,
           failValidate,
@@ -935,8 +991,103 @@ appmap_dir: tmp/appmap
           checkCurrentConfig
         );
       });
+
+      it('yarn getYarnSubprojects for yarn 1, error', async () => {
+        expect.assertions(1);
+        const projectFixture = path.join(fixtureDir, 'javascript', 'yarn_with_subprojects');
+        const installer = new YarnInstaller(projectFixture);
+        jest.spyOn(installer, 'isYarnVersionOne').mockResolvedValue(true);
+        jest.spyOn(commandRunner, 'run').mockResolvedValueOnce({
+          stdout: `yarn workspaces v1.22.19
+error Cannot find the root of your workspace - are you sure you're currently in a workspace?
+info Visit https://yarnpkg.com/en/docs/cli/workspaces for documentation about this command.`,
+          stderr: '',
+        });
+        expect(await getYarnSubprojects(projectFixture, [installer])).toStrictEqual([]);
+      });
+
+      it('yarn getYarnSubprojects for yarn 1, some workspaces', async () => {
+        expect.assertions(1);
+        const projectFixture = path.join(fixtureDir, 'javascript', 'yarn_with_subprojects');
+        const installer = new YarnInstaller(projectFixture);
+        jest.spyOn(installer, 'isYarnVersionOne').mockResolvedValue(true);
+        jest.spyOn(commandRunner, 'run').mockResolvedValueOnce({
+          stdout: `{
+            "subproject_one": {
+              "location": "packages/subproject_one",
+              "workspaceDependencies": [],
+              "mismatchedWorkspaceDependencies": []
+            },
+            "subproject_two": {
+              "location": "packages/subproject_two",
+              "workspaceDependencies": [],
+              "mismatchedWorkspaceDependencies": []
+            }
+          }`,
+          stderr: '',
+        });
+        expect(await getYarnSubprojects(projectFixture, [installer])).toStrictEqual([
+          // it doesn't detect the root workspace
+          {
+            name: 'subproject_one',
+            path: join(projectFixture, 'packages/subproject_one'),
+            availableInstallers: [
+              new YarnInstaller(join(projectFixture, 'packages/subproject_one')),
+            ],
+          },
+          {
+            name: 'subproject_two',
+            path: join(projectFixture, 'packages/subproject_two'),
+            availableInstallers: [
+              new YarnInstaller(join(projectFixture, 'packages/subproject_two')),
+            ],
+          },
+        ]);
+      });
+
+      it('yarn getYarnSubprojects for yarn 2', async () => {
+        expect.assertions(1);
+
+        sinon.restore(); // undo effects of beforeEach, to wrap run() differently
+        sinon
+          .stub(commandRunner, 'run')
+          // yarn workspaces list --json
+          .resolves({
+            stdout: `{"location":".","name":"yarn_with_subprojects"}
+{"location":"packages/subproject_one","name":"subproject_one"}
+{"location":"packages/subproject_two","name":"subproject_two"}`,
+            stderr: '',
+          });
+
+        const projectFixture = path.join(fixtureDir, 'javascript', 'yarn_with_subprojects');
+        const installer = new YarnInstaller(projectFixture);
+        sinon.stub(installer, 'isYarnVersionOne').resolves(false);
+        expect(await getYarnSubprojects(projectFixture, [installer])).toStrictEqual([
+          // it does detect the root workspace
+          {
+            name: 'yarn_with_subprojects',
+            path: projectFixture,
+            availableInstallers: [new YarnInstaller(projectFixture)],
+          },
+          {
+            name: 'subproject_one',
+            path: join(projectFixture, 'packages/subproject_one'),
+            availableInstallers: [
+              new YarnInstaller(join(projectFixture, 'packages/subproject_one')),
+            ],
+          },
+          {
+            name: 'subproject_two',
+            path: join(projectFixture, 'packages/subproject_two'),
+            availableInstallers: [
+              new YarnInstaller(join(projectFixture, 'packages/subproject_two')),
+            ],
+          },
+        ]);
+      });
     });
   });
+
   describe('Varied project configurations', () => {
     beforeEach(() => {
       sinon.stub(commandRunner, 'run').resolves({ stdout: '', stderr: '' });
@@ -1224,6 +1375,130 @@ appmap_dir: tmp/appmap
       });
 
       // Root project exists, should default to not choosing a sub-project.
+      expect(promptStub.getCall(0).args).toMatchObject([
+        {
+          type: 'confirm',
+          message: expect.stringMatching('This directory contains sub-projects'),
+          default: false,
+        },
+      ]);
+    });
+
+    it('yarn detects sub-projects', async () => {
+      expect.assertions(2);
+      const projectFixture = path.join(fixtureDir, 'javascript', 'yarn_with_subprojects');
+      fse.copySync(projectFixture, projectDir);
+
+      let callIdx = 0;
+      sinon.restore(); // undo effects of beforeEach, to wrap run() differently
+      sinon
+        .stub(commandRunner, 'run')
+        // yarn --version
+        .onCall(callIdx++)
+        .resolves({
+          stdout: '3.2.1',
+          stderr: '',
+        })
+        // yarn workspaces list --json
+        .onCall(callIdx++)
+        .resolves({
+          stdout: `{"location":".","name":"yarn_with_subprojects"}
+{"location":"packages/subproject_one","name":"subproject_one"}
+{"location":"packages/subproject_two","name":"subproject_two"}`,
+          stderr: '',
+        })
+        // installing root project
+        .onCall(callIdx++)
+        .resolves({
+          stdout: '{"configuration": { "contents": "" }, "errors": []}',
+          stderr: '',
+        })
+        .onCall(callIdx++)
+        .resolves({
+          stdout: '', // why wasn't '{}' necessary here?
+          stderr: '',
+        })
+        .onCall(callIdx++)
+        .resolves({
+          stdout: '',
+          stderr: '',
+        })
+        // installing first sub-project
+        .onCall(callIdx++)
+        .resolves({
+          stdout: '{"configuration": { "contents": "" }, "errors": []}',
+          stderr: '',
+        })
+        .onCall(callIdx++)
+        .resolves({
+          stdout: '{}',
+          stderr: '',
+        })
+        .onCall(callIdx++)
+        .resolves({
+          stdout: '',
+          stderr: '',
+        })
+        .onCall(callIdx++)
+        .resolves({
+          stdout: '',
+          stderr: '',
+        })
+        .onCall(callIdx++)
+        .resolves({
+          stdout: '',
+          stderr: '',
+        })
+        // installing second sub-project
+        .onCall(callIdx++)
+        .resolves({
+          stdout: '{"configuration": { "contents": "" }, "errors": []}',
+          stderr: '',
+        })
+        .onCall(callIdx++)
+        .resolves({
+          stdout: '{}',
+          stderr: '',
+        })
+        .onCall(callIdx++)
+        .resolves({
+          stdout: '',
+          stderr: '',
+        })
+        .onCall(callIdx++)
+        .resolves({
+          stdout: '',
+          stderr: '',
+        })
+        .onCall(callIdx++)
+        .resolves({
+          stdout: '',
+          stderr: '',
+        })
+        .onCall(callIdx++)
+        .resolves({
+          stdout: '{"configuration": { "contents": "" }, "errors": []}',
+          stderr: '',
+        })
+        .onCall(callIdx++)
+        .resolves({
+          stdout: '{}',
+          stderr: '',
+        });
+
+      const promptStub = sinon.stub(inquirer, 'prompt').resolves({
+        addSubprojects: true,
+        confirm: true,
+        selectedSubprojects: ['yarn_with_subprojects', 'subproject_one', 'subproject_two'],
+      });
+
+      sinon.stub(UI, 'success');
+
+      await invokeCommand(projectDir, (err) => {
+        expect(err).toBeNull();
+      });
+
+      // There's more than one sub-project, should default to not choosing a sub-project.
       expect(promptStub.getCall(0).args).toMatchObject([
         {
           type: 'confirm',